Fixing mistakes and comp problems
