+++ kratos/tests/test_utilities/test_suite.cpp
@@ -8,6 +8,7 @@
 //                   Kratos default license: kratos/license.txt
 //
 //  Main authors:    Carlos A. Roig
+//                   Jordi Cotela Dalmau
 //
 //
 
@@ -16,28 +17,11 @@
 // External includes
 
 // Project includes
-#include "includes/parallel_environment.h"
-
-#include "tests/test_utilities/testing_environment.h"
+#include "tests/test_utilities/test_suite.h"
 
 namespace Kratos::Testing 
 {
-
-void KratosTestEnv::SetUp() {
-    // This is intentionally left empty.
-    // Add here your code if you need to set up some stuff for your tests or debuging
-}
-
-void KratosTestEnv::TearDown() {
-    // This is intentionally left empty.
-    // Add here your code if you need to tear down some stuff for your tests or debuging
-}
-
-KratosTestEnv::KratosTestEnv() {
-    // This is intentionally left empty.
-    // Add here your code if you need to do any operation in initialization or debuging
-}
-
+    
 void KratosCoreFastSuite::SetUp() { 
     // TODO: Control the log level of the tests
     // mCoutBuffer = std::cout.rdbuf();
@@ -51,11 +35,6 @@ void KratosCoreFastSuite::TearDown() {
     // TODO: Control the log level of the tests
     // std::cout.rdbuf(mCoutBuffer);
     // std::cerr.rdbuf(mCerrBuffer); 
-}
-
-DataCommunicator& GetDefaultDataCommunicator()
-{
-    return ParallelEnvironment::GetDefaultDataCommunicator();
 }
 
 } // namespace Kratos::Testing
